Move some of option and cmdline data into BuildContext.
This is to assist in further refactoring of the ksb::Application code to not
require an Internet connection just to construct an object.

This moves the data structures uses to hold cmdline options (to later mask out
Module options), deferred options (to later merge in if not masked), and
selectors chosen into the ksb::BuildContext, where it belongs.  That way, other
classes can refer to the build context instead of having to hold their own
copies.