Rebuild a testing structure.
The test suite for kdesrc-build rather died with the old modularization effort.
Ironically modularization was intended to improve the testing situation; I just
never found the time to complete the work.

I intend to start this work in a separate branch, starting with work to permit
creating relevant kdesrc-build data structures without immediately embarking on
a build, so that we can get better unit tests and (later) integration tests.