non-kde-deps: Add shell handler and metadata for distro dependencies.
This adds the core components for handling per-module dependencies,
including what is for now a "pre-update" phase (for when these updates
would actually be applied) and a dependencies file that is stored with
kdesrc-build for now.

This is separate from the --initial-setup "first run" support since we
cannot rely on YAML metadata at all at that point (which is why
modules/ksb/FirstRun.pm embeds the package list instead).

See issue #9